be more careful with the branch in lexer_recover()
Since it's about the be removed, the offset calculation must be done before
removing it. Then perform the closest stamped token operation as removing the
branch could cause stamped tokens to be removed.

While here, add line number to injected tokens which is of importance while
finding the closest branch during recover.