rework AlignAfterOpenBracket implementation
The current heuristic used to select between aligning with the left
parenthesis or fallback to regular continuation indentation is not
sufficient. Therefore introduce a new document type that given a list of
potential indentations selects the one that minimizes the amount of new
line(s) and line(s) exceeding ColumnLimit.